PeerMarket/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
passlib[bcrypt]
python-jose
python-multipart